Restructured namespaces/modules
