+++ Hook/routes/api/badges.py
